requirements.txt
paho-mqtt==1.6.1
PyYAML>=5.4
requests==2.24.0